Code Fix: Multiple code fixes and error handling.

+++ qtemr/src/sqlcore.cpp
@@ -255,14 +255,26 @@ void sqlCore::closeDataBase()
 
 void sqlCore::processResponse(const QByteArray& response) {
 
+    QJsonDocument jsonResponse = QJsonDocument::fromJson(response);
+    if (!jsonResponse.isArray()){
+        QMessageBox::information(nullptr,"Error","Error downloading update, please try again later.");
+        sendProgress("Download Failed!");
+        emit drugsDatabaseUpdateFinished(false);
+        return;
+    }
+
     if (!query->exec("BEGIN TRANSACTION;")) {
         mDebug() << "Error executing SQL:" << query->lastError().text();
+        emit drugsDatabaseUpdateFinished(false);
         return;
     }
 
 
     if (!query->exec("DROP table IF EXISTS druglist;")) {
         mDebug() << "Error executing SQL:" << query->lastError().text();
+        query->exec("ROLLBACK;");
+        sendProgress("Rolling back!");
+        emit drugsDatabaseUpdateFinished(false);
         return;
     }
 
@@ -282,10 +294,13 @@ void sqlCore::processResponse(const QByteArray& response) {
                     "pharmacology TEXT, "
                     "sold_times TEXT);")) {
         mDebug() << "Error executing SQL:" << query->lastError().text();
+        query->exec("ROLLBACK;");
+        sendProgress("Rolling back!");
+        emit drugsDatabaseUpdateFinished(false);
         return;
     }
 
-    QJsonDocument jsonResponse = QJsonDocument::fromJson(response);
+
     QJsonArray jsonArray = jsonResponse.array();
 
     for (const QJsonValue& value : jsonArray) {
@@ -312,171 +327,178 @@ void sqlCore::processResponse(const QByteArray& response) {
                                   .arg(id, name, arabic, price, active, description, company,
                                        dosage_form, units, barcode, route, pharmacology, sold_times);
 
+        sendProgress("Proccessing database.");
+        qApp->processEvents(QEventLoop::AllEvents);
+
         if (!query->exec(insertQuery)) {
             mDebug() << "Error executing SQL:" << query->lastError().text();
+            query->exec("ROLLBACK;");
+            sendProgress("Rolling back!");
+            emit drugsDatabaseUpdateFinished(false);
             return;
         }
     }
 
+    QStringList quereis = QStringList() <<
+    "UPDATE druglist SET active = UPPER(active);"<<
+    "UPDATE druglist SET description = trim(description);"<<
+    "UPDATE druglist SET description = ltrim(description,'.');"<<
+    "UPDATE druglist SET description = replace(description,'.',' - ');"<<
+    "UPDATE druglist SET description = UPPER(description);"<<
+    "UPDATE druglist SET dosage_form = trim(dosage_form);"<<
+    "UPDATE druglist SET dosage_form = UPPER(dosage_form);"<<
+    "UPDATE druglist SET company = trim(company);"<<
+    "UPDATE druglist SET company = ltrim(company,'.');"<<
+    "UPDATE druglist SET company = ltrim(company,'\"');"<<
+    "UPDATE druglist SET company = UPPER(company);"<<
+    "UPDATE druglist SET name = TRIM(name);"<<
+    "UPDATE druglist SET name = UPPER(name);"<<
+    "UPDATE druglist SET name = replace(name,'\"','');"<<
+    "UPDATE druglist SET name = replace(name,' MG','MG');"<<
+    "UPDATE druglist SET name = replace(name,' MCG','MCG');"<<
+    "UPDATE druglist SET name = replace(name,' ML','ML');"<<
+    "UPDATE druglist SET name = replace(name,' MG','MG');"<<
+    "UPDATE druglist SET name = replace(name,' MCG','MCG');"<<
+    "UPDATE druglist SET name = replace(name,' ML','ML');"<<
+    "UPDATE druglist SET name = replace(name,' GRAM','GRAM');"<<
+    "UPDATE druglist SET name = replace(name,' GM','GM');"<<
+    "UPDATE druglist SET name = replace(name,' GRAM','GRAM');"<<
+    "UPDATE druglist SET name = replace(name,' GM','GM');"<<
+    "UPDATE druglist SET name = replace(name,'M.I.U.','MIU');"<<
+    "UPDATE druglist SET name = replace(name,'M.I.U','MIU');"<<
+    "UPDATE druglist SET name = replace(name,'M.IU','MIU');"<<
+    "UPDATE druglist SET name = replace(name,'  MIU','MIU');"<<
+    "UPDATE druglist SET name = replace(name,' MIU','MIU');"<<
+    "UPDATE druglist SET name = replace(name,' IU-','IU');"<<
+    "UPDATE druglist SET name = replace(name,' IU','IU');"<<
+    "UPDATE druglist SET name = replace(name,' I.U.','IU');"<<
+    "UPDATE druglist SET name = replace(name,'I.U.','IU');"<<
+    "UPDATE druglist SET name = replace(name,'TABLET','TAB');"<<
+    "UPDATE druglist SET name = replace(name,'CAPSULE','CAP');"<<
+    "UPDATE druglist SET name = replace(name,'TABS','TAB');"<<
+    "UPDATE druglist SET name = replace(name,'CAPS','CAP');"<<
+    "UPDATE druglist SET name = replace(name,'...','.');"<<
+    "UPDATE druglist SET name = replace(name,'..','.');"<<
+    "UPDATE druglist SET name = rtrim(name,'.');"<<
+    "UPDATE druglist SET name = replace(name,'AMPS','AMP');"<<
+    "UPDATE druglist SET name = replace(name,'AMPSS','AMP');"<<
+    "UPDATE druglist SET name = replace(name,'F.C.TAB','F.C. TAB');"<<
+    "UPDATE druglist SET name = replace(name,'BONE CARE BONE CARE 0.5MCG 30 CAP. 30 CAP','BONE CARE 0.5MCG 30 CAP');"<<
+    "UPDATE druglist SET name = replace(name,'DISTAB','DIS. TAB');"<<
+    "UPDATE druglist SET name = replace(name,'DISTABLETS','DIS. TAB');"<<
+    "UPDATE druglist SET name = replace(name,'-',' ');"<<
+    "UPDATE druglist SET name = replace(name,' TABLES','TAB');"<<
+    "UPDATE druglist SET name = replace(name,'IUVIAL','IU VIAL');"<<
+    "UPDATE druglist SET name = replace(name,'.VIAL',' VIAL');"<<
+    "UPDATE druglist SET name = replace(name,'.AMP',' AMP');"<<
+    "UPDATE druglist SET name = replace(name,'.TAB',' TAB');"<<
+    "UPDATE druglist SET name = replace(name,'.CAP',' CAP');"<<
+    "UPDATE druglist SET name = replace(name,'S.R','SR ');"<<
+    "UPDATE druglist SET name = replace(name,'M.R.','MR');"<<
+    "UPDATE druglist SET name = replace(name,' M.R','MR');"<<
+    "UPDATE druglist SET name = replace(name,'I.V','IV');"<<
+    "UPDATE druglist SET name = replace(name,'F.C.','FC');"<<
+    "UPDATE druglist SET name = replace(name,'F.C','FC');"<<
+    "UPDATE druglist SET name = replace(name,'S.C.','SC');"<<
+    "UPDATE druglist SET name = replace(name,'S.C','SC');"<<
+    "UPDATE druglist SET name = replace(name,'IM.','IM ');"<<
+    "UPDATE druglist SET name = replace(name,'I.M.','IM ');"<<
+    "UPDATE druglist SET name = replace(name,'I.M','IM');"<<
+    "UPDATE druglist SET name = replace(name,'I.V.','IV ');"<<
+    "UPDATE druglist SET name = replace(name,'IV.','IV ');"<<
+    "UPDATE druglist SET name = replace(name,'VIAL.','VIAL ');"<<
+    "UPDATE druglist SET name = replace(name,'VIALS','VIAL ');"<<
+    "UPDATE druglist SET name = replace(name,'AMP.','AMP ');"<<
+    "UPDATE druglist SET name = replace(name,'CAP.','CAP ');"<<
+    "UPDATE druglist SET name = replace(name,'TAB.','TAB ');"<<
+    "UPDATE druglist SET name = replace(name,'DROPS.','DROPS');"<<
+    "UPDATE druglist SET name = replace(name,'PRE FILLED','PREFILLED');"<<
+    "UPDATE druglist SET name = replace(name,' /IM','/IM');"<<
+    "UPDATE druglist SET name = replace(name,' /IV','/IV');"<<
+    "UPDATE druglist SET name = replace(name,'IV /','IV/');"<<
+    "UPDATE druglist SET name = replace(name,'IM /','IM/');"<<
+    "UPDATE druglist SET name = replace(name,' /SC','/SC');"<<
+    "UPDATE druglist SET name = replace(name,'   ',' ');"<<
+    "UPDATE druglist SET name = replace(name,'  ',' ');"<<
+    "UPDATE druglist SET name = replace(name,'IV / IM','IV/IM');"<<
+    "UPDATE druglist SET name = replace(name,'1 2 3 (ONE TWO THREE)','123');"<<
+    "UPDATE druglist SET name = replace(name,'1 2 3','123');"<<
+    "UPDATE druglist SET name = RTRIM(name,'$$');"<<
+    "UPDATE druglist SET name = replace(name,'ZORA C 20/LOZENGES','ZORA C 20 LOZENGES');"<<
+    "DELETE FROM druglist WHERE name='ZORA C 20 LOZENGE';"<<
+    "UPDATE druglist SET name = TRIM(name);"<<
+    "UPDATE druglist SET name='CIPROCORT EAR DROPS 10ML' WHERE name = 'CIPROCORT OTIC DROPS 10ML';"<<
+    "UPDATE druglist SET dosage_form='EAR DROPS' WHERE name = 'CIPROCORT EAR DROPS 10ML';"<<
+    "UPDATE druglist SET description = 'MULTIVITAMIN' WHERE name = 'REGNADEX 30 TAB';"<<
+    "UPDATE druglist SET dosage_form = 'CAPSULE' WHERE dosage_form = 'CAPS';"<<
+    "UPDATE druglist SET dosage_form = 'CREAM' WHERE dosage_form = 'CRE';"<<
+    "UPDATE druglist SET dosage_form = 'POWDER' WHERE dosage_form = 'POWER';"<<
+    "UPDATE druglist SET dosage_form = 'TABLET' WHERE dosage_form = 'TABS.' OR dosage_form = 'TABLETS' ;"<<
+    "UPDATE druglist SET active = 'UNSPECIFIED' WHERE TRIM(active) = '';"<<
+    QString("UPDATE metadata SET value=%1 WHERE var='version'").arg(QDate::currentDate().toString("yyMMdd"));
+
+    sendProgress("Fixing Database.");
+    qApp->processEvents(QEventLoop::AllEvents);
+
+    int fixes = quereis.count();
+    int fix = 0;
+    for(const auto &q :quereis){
+        sendProgress(QString("DB fix %1 of %2").arg(fix).arg(fixes));
+        qApp->processEvents(QEventLoop::AllEvents);
+        if(!query->exec(q)){
+            query->exec("ROLLBACK;");
+            sendProgress("Rolling back!");
+            QMessageBox::information(nullptr,"Error","updating drugs database, please try again later.");
+            emit drugsDatabaseUpdateFinished(false);
+            return;
+        }
+        fix++;
+    }
+    //SELECT name, COUNT(*) c FROM druglist GROUP BY name HAVING c > 1 AND TRIM(pharmacology) != '';
+
     if (!query->exec("COMMIT;")) {
         mDebug() << "Error executing SQL:" << query->lastError().text();
     }
+    sendProgress("Done!");
+    emit drugsDatabaseUpdateFinished(true);
+}
 
-    query->exec("UPDATE druglist SET active = UPPER(active);");
-    query->exec("UPDATE druglist SET description = trim(description);");
-    query->exec("UPDATE druglist SET description = ltrim(description,'.');");
-    query->exec("UPDATE druglist SET description = replace(description,'.',' - ');");
-    query->exec("UPDATE druglist SET description = UPPER(description);");
-    query->exec("UPDATE druglist SET dosage_form = trim(dosage_form);");
-    query->exec("UPDATE druglist SET dosage_form = UPPER(dosage_form);");
-    query->exec("UPDATE druglist SET company = trim(company);");
-    query->exec("UPDATE druglist SET company = ltrim(company,'.');");
-    query->exec("UPDATE druglist SET company = ltrim(company,'\"');");
-    query->exec("UPDATE druglist SET company = UPPER(company);");
+void sqlCore::sendProgress(const QString &status)
+{
+    progress(QString("<b>%1</b>").arg(status));
+}
 
-    query->exec("UPDATE druglist SET name = TRIM(name);");
-    query->exec("UPDATE druglist SET name = UPPER(name);");
-    query->exec("UPDATE druglist SET name = replace(name,'\"','');");
-
-    query->exec("UPDATE druglist SET name = replace(name,' MG','MG');");
-    query->exec("UPDATE druglist SET name = replace(name,' MCG','MCG');");
-    query->exec("UPDATE druglist SET name = replace(name,' ML','ML');");
-
-    query->exec("UPDATE druglist SET name = replace(name,' MG','MG');");
-    query->exec("UPDATE druglist SET name = replace(name,' MCG','MCG');");
-    query->exec("UPDATE druglist SET name = replace(name,' ML','ML');");
-
-    query->exec("UPDATE druglist SET name = replace(name,' GRAM','GRAM');");
-    query->exec("UPDATE druglist SET name = replace(name,' GM','GM');");
-
-    query->exec("UPDATE druglist SET name = replace(name,' GRAM','GRAM');");
-    query->exec("UPDATE druglist SET name = replace(name,' GM','GM');");
-
-
-    query->exec("UPDATE druglist SET name = replace(name,'M.I.U.','MIU');");
-    query->exec("UPDATE druglist SET name = replace(name,'M.I.U','MIU');");
-    query->exec("UPDATE druglist SET name = replace(name,'M.IU','MIU');");
-    query->exec("UPDATE druglist SET name = replace(name,'  MIU','MIU');");
-    query->exec("UPDATE druglist SET name = replace(name,' MIU','MIU');");
-    query->exec("UPDATE druglist SET name = replace(name,' IU-','IU');");
-    query->exec("UPDATE druglist SET name = replace(name,' IU','IU');");
-    query->exec("UPDATE druglist SET name = replace(name,' I.U.','IU');");
-    query->exec("UPDATE druglist SET name = replace(name,'I.U.','IU');");
-
-
-    query->exec("UPDATE druglist SET name = replace(name,'TABLET','TAB');");
-    query->exec("UPDATE druglist SET name = replace(name,'CAPSULE','CAP');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'TABS','TAB');");
-    query->exec("UPDATE druglist SET name = replace(name,'CAPS','CAP');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'...','.');");
-    query->exec("UPDATE druglist SET name = replace(name,'..','.');");
-    query->exec("UPDATE druglist SET name = rtrim(name,'.');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'AMPS','AMP');");
-    query->exec("UPDATE druglist SET name = replace(name,'AMPSS','AMP');");
-    query->exec("UPDATE druglist SET name = replace(name,'F.C.TAB','F.C. TAB');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'BONE CARE BONE CARE 0.5MCG 30 CAP. 30 CAP','BONE CARE 0.5MCG 30 CAP');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'DISTAB','DIS. TAB');");
-    query->exec("UPDATE druglist SET name = replace(name,'DISTABLETS','DIS. TAB');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'-',' ');");
-    query->exec("UPDATE druglist SET name = replace(name,' TABLES','TAB');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'IUVIAL','IU VIAL');");
-    query->exec("UPDATE druglist SET name = replace(name,'.VIAL',' VIAL');");
-    query->exec("UPDATE druglist SET name = replace(name,'.AMP',' AMP');");
-    query->exec("UPDATE druglist SET name = replace(name,'.TAB',' TAB');");
-    query->exec("UPDATE druglist SET name = replace(name,'.CAP',' CAP');");
-    query->exec("UPDATE druglist SET name = replace(name,'S.R','SR ');");
-    query->exec("UPDATE druglist SET name = replace(name,'M.R.','MR');");
-    query->exec("UPDATE druglist SET name = replace(name,' M.R','MR');");
-    query->exec("UPDATE druglist SET name = replace(name,'I.V','IV');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'F.C.','FC');");
-    query->exec("UPDATE druglist SET name = replace(name,'F.C','FC');");
-    query->exec("UPDATE druglist SET name = replace(name,'S.C.','SC');");
-    query->exec("UPDATE druglist SET name = replace(name,'S.C','SC');");
-
-
-    query->exec("UPDATE druglist SET name = replace(name,'IM.','IM ');");
-    query->exec("UPDATE druglist SET name = replace(name,'I.M.','IM ');");
-    query->exec("UPDATE druglist SET name = replace(name,'I.M','IM');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'I.V.','IV ');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'IV.','IV ');");
-    query->exec("UPDATE druglist SET name = replace(name,'VIAL.','VIAL ');");
-    query->exec("UPDATE druglist SET name = replace(name,'VIALS','VIAL ');");
-    query->exec("UPDATE druglist SET name = replace(name,'AMP.','AMP ');");
-    query->exec("UPDATE druglist SET name = replace(name,'CAP.','CAP ');");
-    query->exec("UPDATE druglist SET name = replace(name,'TAB.','TAB ');");
-    query->exec("UPDATE druglist SET name = replace(name,'DROPS.','DROPS');");
-    query->exec("UPDATE druglist SET name = replace(name,'PRE FILLED','PREFILLED');");
-
-    query->exec("UPDATE druglist SET name = replace(name,' /IM','/IM');");
-    query->exec("UPDATE druglist SET name = replace(name,' /IV','/IV');");
-    query->exec("UPDATE druglist SET name = replace(name,'IV /','IV/');");
-    query->exec("UPDATE druglist SET name = replace(name,'IM /','IM/');");
-
-    query->exec("UPDATE druglist SET name = replace(name,' /SC','/SC');");
-
-
-
-    query->exec("UPDATE druglist SET name = replace(name,'   ',' ');");
-    query->exec("UPDATE druglist SET name = replace(name,'  ',' ');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'IV / IM','IV/IM');");
-
-    query->exec("UPDATE druglist SET name = replace(name,'1 2 3 (ONE TWO THREE)','123');");
-    query->exec("UPDATE druglist SET name = replace(name,'1 2 3','123');");
-
-    query->exec("UPDATE druglist SET name = RTRIM(name,'$$');");
-    query->exec("UPDATE druglist SET name = replace(name,'ZORA C 20/LOZENGES','ZORA C 20 LOZENGES');");
-    query->exec("DELETE FROM druglist WHERE name='ZORA C 20 LOZENGE';");
-
-    query->exec("UPDATE druglist SET name = TRIM(name);");
-
-    query->exec("UPDATE druglist SET name='CIPROCORT EAR DROPS 10ML' WHERE name = 'CIPROCORT OTIC DROPS 10ML';");
-    query->exec("UPDATE druglist SET dosage_form='EAR DROPS' WHERE name = 'CIPROCORT EAR DROPS 10ML';");
-    query->exec("UPDATE druglist SET description = 'MULTIVITAMIN' WHERE name = 'REGNADEX 30 TAB';");
-
-    query->exec("UPDATE druglist SET dosage_form = 'CAPSULE' WHERE dosage_form = 'CAPS';");
-    query->exec("UPDATE druglist SET dosage_form = 'CREAM' WHERE dosage_form = 'CRE';");
-    query->exec("UPDATE druglist SET dosage_form = 'POWDER' WHERE dosage_form = 'POWER';");
-    query->exec("UPDATE druglist SET dosage_form = 'TABLET' WHERE dosage_form = 'TABS.' OR dosage_form = 'TABLETS' ;");
-
-
-
-    query->exec("UPDATE druglist SET active = 'UNSPECIFIED' WHERE TRIM(active) = '';");
-    query->exec(QString("UPDATE metadata SET value=%1 WHERE var='version'").arg(QDate::currentDate().toString("yyMMdd")));
-    //SELECT name, COUNT(*) c FROM druglist GROUP BY name HAVING c > 1 AND TRIM(pharmacology) != '';
-
-    query->exec("COMMIT;");
-
-
-    emit drugsDatabaseUpdateFinished();
+void sqlCore::downloadProgress(qint64 a, qint64 b)
+{
+    Q_UNUSED(b);
+    progress(QString("<b>Got %1 KiB.</b>").arg(a/1000));
 }
 
 void sqlCore::updateDrugsDatabase()
 {
-
     QNetworkRequest request(QUrl("https://dwaprices.com/api_dr88g/index.php"));
     request.setHeader(QNetworkRequest::ContentTypeHeader, "application/x-www-form-urlencoded");
 
     QObject::connect(&networkManager, &QNetworkAccessManager::finished, [&](QNetworkReply *reply) {
+
         if (reply->error() == QNetworkReply::NoError) {
             QByteArray response = reply->readAll();
+            progress(QString("<b>Download Finished.</b>"));
+            qApp->processEvents(QEventLoop::AllEvents);
             processResponse(response);
         } else {
             qDebug() << "Error:" << reply->errorString();
         }
         reply->deleteLater();
     });
-    networkManager.post(request, "updatesqlite=1");
+
+    QNetworkReply *reply = networkManager.post(request, "updatesqlite=1");
+    connect(reply,&QNetworkReply::downloadProgress,this,&sqlCore::downloadProgress);
+    connect(reply,&QNetworkReply::errorOccurred,[=](){
+        reply->abort();
+        reply->deleteLater();
+        drugsDatabaseUpdateFinished(false);
+    });
 }
